requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
asttokens=2.2.1=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blas=1.0=mkl
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
brotlipy=0.7.0=py310h2bbff1b_1002
bzip2=1.0.8=he774522_0
ca-certificates=2023.01.10=haa95532_0
certifi=2022.12.7=py310haa95532_0
cffi=1.15.1=py310h2bbff1b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
colorama=0.4.6=pyhd8ed1ab_0
contourpy=1.0.5=py310h59b6b97_0
cryptography=38.0.4=py310h21b164f_0
cuda=11.7.1=0
cuda-cccl=11.7.91=0
cuda-command-line-tools=11.7.1=0
cuda-compiler=11.7.1=0
cuda-cudart=11.7.99=0
cuda-cudart-dev=11.7.99=0
cuda-cuobjdump=11.7.91=0
cuda-cupti=11.7.101=0
cuda-cuxxfilt=11.7.91=0
cuda-demo-suite=12.0.140=0
cuda-documentation=12.0.140=0
cuda-libraries=11.7.1=0
cuda-libraries-dev=11.7.1=0
cuda-memcheck=11.8.86=0
cuda-nsight-compute=12.0.1=0
cuda-nvcc=11.7.99=0
cuda-nvdisasm=12.0.140=0
cuda-nvml-dev=11.7.91=0
cuda-nvprof=12.0.146=0
cuda-nvprune=11.7.91=0
cuda-nvrtc=11.7.99=0
cuda-nvrtc-dev=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-nvvp=12.0.146=0
cuda-runtime=11.7.1=0
cuda-sanitizer-api=12.0.140=0
cuda-toolkit=11.7.1=0
cuda-tools=11.7.1=0
cuda-visual-tools=11.7.1=0
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.5.1=py310hd77b12b_0
decorator=5.1.1=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
fftw=3.3.9=h2bbff1b_1
flit-core=3.6.0=pyhd3eb1b0_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=ha860e81_0
future=0.18.3=pypi_0
glib=2.69.1=h5dc1a3c_2
gst-plugins-base=1.18.5=h9e645db_0
gstreamer=1.18.5=hd78058f_0
h5py=3.8.0=pypi_0
icc_rt=2022.1.0=h6049295_2
icu=58.2=ha925a31_3
idna=3.4=py310haa95532_0
importlib-metadata=6.0.0=pyha770c72_0
importlib_metadata=6.0.0=hd8ed1ab_0
intel-openmp=2021.4.0=haa95532_3556
ipykernel=6.15.0=pyh025b116_0
ipython=8.10.0=pyh08f2357_0
jedi=0.18.2=pyhd8ed1ab_0
joblib=1.1.1=py310haa95532_0
jpeg=9e=h2bbff1b_0
jupyter_client=8.0.2=pyhd8ed1ab_0
jupyter_core=5.2.0=py310h5588dad_0
kiwisolver=1.4.4=py310hd77b12b_0
lerc=3.0=hd77b12b_0
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libclang=12.0.0=default_h627e005_2
libcublas=11.10.3.66=0
libcublas-dev=11.10.3.66=0
libcufft=10.7.2.124=0
libcufft-dev=10.7.2.124=0
libcurand=10.3.1.124=0
libcurand-dev=10.3.1.124=0
libcusolver=11.4.0.1=0
libcusolver-dev=11.4.0.1=0
libcusparse=11.7.4.91=0
libcusparse-dev=11.7.4.91=0
libdeflate=1.8=h2bbff1b_5
libffi=3.4.2=hd77b12b_6
libiconv=1.16=h2bbff1b_2
libnpp=11.7.4.75=0
libnpp-dev=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libnvjpeg-dev=11.8.0.2=0
libogg=1.3.5=h2bbff1b_1
libpng=1.6.37=h2a8f88b_0
libsodium=1.0.18=h8d14728_1
libtiff=4.5.0=h6c2663c_1
libuv=1.40.0=he774522_0
libvorbis=1.3.7=he774522_0
libwebp=1.2.4=h2bbff1b_0
libwebp-base=1.2.4=h2bbff1b_0
libxml2=2.9.14=h0ad7f3c_0
libxslt=1.1.35=h2bbff1b_0
lz4-c=1.9.4=h2bbff1b_0
mat4py=0.5.0=pypi_0
matplotlib=3.6.2=py310haa95532_0
matplotlib-base=3.6.2=py310h1094b8e_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py310h2bbff1b_0
mkl_fft=1.3.1=py310ha0764ea_0
mkl_random=1.2.2=py310h4ed8f06_0
munkres=1.1.4=py_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
nsight-compute=2022.4.1.6=0
numpy=1.23.5=py310h60c9a35_0
numpy-base=1.23.5=py310h04254f7_0
openssl=1.1.1t=h2bbff1b_0
packaging=23.0=pyhd8ed1ab_0
parso=0.8.3=pyhd8ed1ab_0
pcre=8.45=hd77b12b_0
pickleshare=0.7.5=py_1003
pillow=9.3.0=py310hd77b12b_2
pip=23.0=pyhd8ed1ab_0
platformdirs=2.5.2=py310haa95532_0
ply=3.11=py310haa95532_0
prompt-toolkit=3.0.36=pyha770c72_0
psutil=5.9.0=py310h2bbff1b_0
pure_eval=0.2.2=pyhd8ed1ab_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.14.0=pyhd8ed1ab_0
pymatreader=0.0.30=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py310haa95532_0
pyqt=5.15.7=py310hd77b12b_0
pyqt5-sip=12.11.0=py310hd77b12b_0
pysocks=1.7.1=py310haa95532_0
python=3.10.9=h966fe2a_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.10=2_cp310
pytorch=1.13.1=py3.10_cuda11.7_cudnn8_0
pytorch-cuda=11.7=h67b0de4_1
pytorch-mutex=1.0=cuda
pywin32=305=py310h2bbff1b_0
pyzmq=23.2.0=py310hd77b12b_0
qt-main=5.15.2=he8e5bd7_7
qt-webengine=5.15.9=hb9a9bb5_5
qtwebkit=5.212=h3ad3cdb_4
requests=2.28.1=py310haa95532_0
scikit-learn=1.2.0=py310hd77b12b_1
scipy=1.10.0=pypi_0
setuptools=65.6.3=py310haa95532_0
sip=6.6.2=py310hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
sklearn=0.0.post1=pypi_0
sqlite=3.40.1=h2bbff1b_0
stack_data=0.6.2=pyhd8ed1ab_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h2bbff1b_0
toml=0.10.2=pyhd3eb1b0_0
torchaudio=0.13.1=pypi_0
torchvision=0.14.1=pypi_0
tornado=6.2=py310he2412df_0
tqdm=4.64.1=pypi_0
traitlets=5.9.0=pyhd8ed1ab_0
typing_extensions=4.4.0=py310haa95532_0
tzdata=2022g=h04d1e81_0
urllib3=1.26.14=py310haa95532_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.6=pyhd8ed1ab_0
wheel=0.37.1=pyhd3eb1b0_0
win_inet_pton=1.1.0=py310haa95532_0
wincertstore=0.2=py310haa95532_2
xmltodict=0.13.0=pypi_0
xz=5.2.10=h8cc25b3_1
zeromq=4.3.4=h0e60522_1
zipp=3.13.0=pyhd8ed1ab_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.2=h19a0ad4_0
